:bug: fix: addition of if statement if build folder still does not exist

+++ build.sh
@@ -3,17 +3,17 @@
 #
 # Given parameters are passed over to CMake.
 # Examples:
-#    * ./build_all.sh -DCMAKE_BUILD_TYPE=Debug
-#    * ./build_all.sh VERBOSE=1
-#
-# Written by Tiffany Huang, 12/14/2016
+#    * ./build.sh -DCMAKE_BUILD_TYPE=Debug
+#    * ./build.sh VERBOSE=1
 #
 
-# Go into the directory where this bash script is contained.
-cd `dirname $0`
+cd `dirname $0` # where this bash script is contained.
 
-rm -rf build
-echo Got rid of old build!
+if [ -d "build" ]; then
+  rm -rf build
+  echo Got rid of old build!
+fi
+
 echo Compiling new code
 mkdir -p build
 cd build
